Moved utils to difi_utils to permit interacting with these classes as a package and updated all references
