+++ DIFI_Validator/drx.py
@@ -46,13 +46,13 @@ import yaml
 import csv
 from scapy.all import *
 
-from utils.difi_constants import *
-from utils.custom_error_types import *
-from utils.file_writing import *
-from utils.noncompliant_class import DifiInfo
-from utils.difi_data_packet_class import DifiDataPacket
-from utils.difi_context_packet_class import DifiStandardContextPacket
-from utils.difi_version_packet_class import DifiVersionContextPacket
+from difi_utils.difi_constants import *
+from difi_utils.custom_error_types import *
+from difi_utils.file_writing import *
+from difi_utils.noncompliant_class import DifiInfo
+from difi_utils.difi_data_packet_class import DifiDataPacket
+from difi_utils.difi_context_packet_class import DifiStandardContextPacket
+from difi_utils.difi_version_packet_class import DifiVersionContextPacket
 
 ##########
 # Settings
